use DBSpendUpdateWriter class for managing DB spend updates
