+++ tests/proxy_unit_tests/test_unit_test_proxy_hooks.py
@@ -5,6 +5,7 @@ from unittest.mock import Mock, patch, AsyncMock
 import pytest
 from fastapi import Request
 from litellm.proxy.utils import _get_redoc_url, _get_docs_url
+from datetime import datetime
 
 sys.path.insert(0, os.path.abspath("../.."))
 import litellm
@@ -22,16 +23,20 @@ async def test_disable_spend_logs():
     with patch("litellm.proxy.proxy_server.disable_spend_logs", True), patch(
         "litellm.proxy.proxy_server.prisma_client", mock_prisma_client
     ):
-        from litellm.proxy.proxy_server import update_database
+        from litellm.proxy.db.db_spend_update_writer import DBSpendUpdateWriter
 
         # Call update_database with disable_spend_logs=True
-        await update_database(
+        await DBSpendUpdateWriter.update_database(
             token="fake-token",
             response_cost=0.1,
             user_id="user123",
             completion_response=None,
-            start_time="2024-01-01",
-            end_time="2024-01-01",
+            start_time=datetime.now(),
+            end_time=datetime.now(),
+            end_user_id="end_user_id",
+            team_id="team_id",
+            org_id="org_id",
+            kwargs={},
         )
         # Verify no spend logs were added
         assert len(mock_prisma_client.spend_log_transactions) == 0
